requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
pandas==2.2.2
python-dotenv==1.0.1
